refactor : ApiResponse -> 공용 패키지로 이동
