+++ src/main/java/com/zerobase/hoops/gameUsers/controller/GameUserController.java
@@ -1,6 +1,7 @@
 
 package com.zerobase.hoops.gameUsers.controller;
 
+import com.zerobase.hoops.commonResponse.ApiResponse;
 import com.zerobase.hoops.gameCreator.type.CityName;
 import com.zerobase.hoops.gameCreator.type.FieldStatus;
 import com.zerobase.hoops.gameCreator.type.Gender;
@@ -92,9 +93,9 @@ public class GameUserController {
 
   @PreAuthorize("hasRole('USER')")
   @PostMapping("/manner-point")
-  public ResponseEntity<ApiResponse2> saveMannerPoint(
+  public ResponseEntity<ApiResponse> saveMannerPoint(
       @RequestBody @Valid MannerPointDto request) {
     gameUserService.saveMannerPoint(request);
-    return ResponseEntity.ok().body(new ApiResponse2("매너점수평가", "Success"));
+    return ResponseEntity.ok().body(new ApiResponse("매너점수평가", "Success"));
   }
 }
